+++ jupyter_io/__init__.py
@@ -1,7 +1,7 @@
 __all__ = [
     # submodules
+    "io",
     "v0",
-    "save",
     # aliases
     "to_html",
     "in_notebook",
@@ -13,14 +13,14 @@ __version__ = "1.0.0"
 
 
 # submodules
+from . import io
 from . import v0
-from . import save
 
 
 # aliases
+from .io import to_html, in_notebook
 from .v0 import (
     savefile_in_notebook,
     savefig_in_notebook,
     savetable_in_notebook,
 )
-from .save import to_html, in_notebook
